+++ banking/src/main/java/com/project88/banking/service/ITransactionService.java
@@ -8,5 +8,5 @@ import com.project88.banking.dto.filter.TransactionFilter;
 
 public interface ITransactionService {
 
-    Page<TransactionHistoryDTO> getTransaction(short userID, Pageable pageable, TransactionFilter filter);
+    Page<TransactionHistoryDTO> getTransaction(long userID, Pageable pageable, TransactionFilter filter);
 }
